refactor: rss와 naver의 구조 통일
